test(e2e): add new test runner
